Remove _v2 from metadata
