+++ dcpy/connectors/socrata/metadata.py
@@ -5,7 +5,7 @@ from typing import Any
 from dcpy.connectors.socrata import publish as pub
 from dcpy.connectors.socrata import metadata
 
-import dcpy.models.product.dataset.metadata_v2 as md
+import dcpy.models.product.dataset.metadata as md
 from dcpy.utils.logging import logger
 
 soc_types_to_dcp_types = {
