+++ dcpy/models/product/data_dictionary.py
@@ -1,4 +1,4 @@
-from .dataset.metadata_v2 import CustomizableBase
+from .dataset.metadata import CustomizableBase
 from dcpy.models.base import TemplatedYamlReader
