+++ dcpy/test/models/product/dataset/test_text_cleaning.py
@@ -1,4 +1,4 @@
-import dcpy.models.product.dataset.metadata_v2 as md_v2
+import dcpy.models.product.dataset.metadata as md
 
 
 def test_description_cleaning():
@@ -16,4 +16,4 @@ def test_description_cleaning():
 
     Two Newlines Above (Should be Preserved)
     Newline below (Should be removed)"""
-    assert md_v2.normalize_text(messy_description) == clean_description
+    assert md.normalize_text(messy_description) == clean_description
